% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{build_master_unit}
\alias{build_master_unit}
\title{Monta Tabela Mestra}
\usage{
build_master_unit(tabela)
}
\arguments{
\item{tabela}{objeto de classe \code{tabela} para o qual construir uma mestra}
}
\value{
tabela mestra
}
\description{
Constroi uma tabela mestra associando particoes as entidades contendo aquele trecho
}
\details{
A tabela mestra nada mais e do que um data.table simples com duas ou mais colunas. Uma destas se
chamara \code{tabela} e contera o nome de um arquivo que componha a tabela abstrata. As demais
colunas sao nomeadas pelos campos de particionamento e conterao, a cada linha, o elemento daquele
campo contido na particao associada

O uso de tabelas mestras nao e estritamente necessario. Pela nomenclatura de particionamento,
seria possivel simplesmente procurar nos nomes de arquivos quais sao os que devem ser lidos. Isso
funciona tranquilamente quando sao mocks locais, mas no s3 demandaria listagem dos buckets toda
vez que uma leitura particionada ocorresse, o que nao e ideal.
}
\examples{

# para a tabela exemplo 'assimilacao' particionada por 'codigo' e 'dia_assimilacao'
arq <- system.file("extdata/cpart_parquet/ass/schema.json", package = "dbrenovaveis")
tab <- dbrenovaveis:::schema2tabela(arq)

mestra <- dbrenovaveis:::build_master_unit(tab)

\dontrun{
print(mestra)
#>       codigo dia_assimilacao                                         tabela
#> 1: AVERMELHA               1 assimilacao-codigo=AVERMELHA-dia_assimilacao=1
#> 2: AVERMELHA               2 assimilacao-codigo=AVERMELHA-dia_assimilacao=2
#> 3:   BAIXOIG               1   assimilacao-codigo=BAIXOIG-dia_assimilacao=1
#> 4:   BAIXOIG               2   assimilacao-codigo=BAIXOIG-dia_assimilacao=2
}

}
