% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-mock.r
\name{checa_particao}
\alias{checa_particao}
\title{Checa Existencia De Particoes Locais}
\usage{
checa_particao(conexao, query)
}
\arguments{
\item{conexao}{objeto de conexao ao banco retornado por \code{\link{conectabanco}}}

\item{query}{lista detalhando a query como retornado por \code{\link{parseargs}}}
}
\value{
booleano indicando se a tabela e particionada ou nao
}
\description{
Avalia se uma tabela local corresponde a conjunto de particoes ou nao
}
