% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfromdb.r
\name{collate_query}
\alias{collate_query}
\title{Compoe Query Completa}
\usage{
collate_query(query)
}
\arguments{
\item{query}{lista de trechos SELECT, FROM e WHERE como retornado por \code{\link{parseargs}}}
}
\value{
string com a query completa
}
\description{
Auxiliar para transformar a lista de query por trecho numa unica string
}
