% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.r
\name{compoe_schema}
\alias{compoe_schema}
\title{Le E Interpreta Schema De Banco}
\usage{
compoe_schema(arq_schema_banco, schema_banco = NULL)
}
\arguments{
\item{arq_schema_banco}{caminho de um arquivo \code{schema.json} para ler}

\item{schema_banco}{lista contendo um arquivo ja lido. Veja Detalhes}
}
\value{
lista contendo o schema do banco processado
}
\description{
Realiza a leitura de um \code{schema.json} para banco, valida e sanitariza seus conteudos se
necessario
}
\details{
Uma das principais sanitarizacoes que esta funcao realiza e o controle de caminhos relativos. O
recomendado e que todas as chaves \code{uri} em arquivos schema sejam sempre caminhos absolutos.
Por outro lado, para fins de testagem do pacote isto nao funciona pois nao e possivel saber de
antemao aonde o pacote sera instalado.

Quando a funcao e chamada com \code{arq_schema_banco}, e possivel identificar se caminhos
relativos existem e converte-los de acordo caso o argumento seja um caminho absoluto por si; do
contrario nao sera possivel resolver os caminhos relativos. Caso seja chamada com
\code{schema_banco}  e esta lista nao contiver um elemento \code{uri}, nao sera possivel
sanitarizar caminhos relativos
}
