% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{conectabanco}
\alias{conectabanco}
\title{Conexao Com Bancos}
\usage{
conectabanco(usuario, banco)
}
\arguments{
\item{usuario}{a tag associada a credenciais previamente regristradas com
\code{\link{registra_credenciais}}}

\item{banco}{a tag associada a informacoes de banco previamente regristradas com
\code{\link{registra_banco}}}
}
\value{
objeto de conexao ao banco. Veja \code{\link[DBI]{dbConnect}} para mais detalhes
}
\description{
Gera a conexao com o banco \code{banco} utilizando as credenciais de \code{usuario}
}
\details{
Esta funcao e um simples wrapper para facilitacao da conexao com bancos de dados, utilizando
as credenciais previamente registradas e salvas pelas funcoes apropriadas.
}
\examples{

# supondo usuario 'user' e banco 'db' ja registrados

\dontrun{
# conecta ao banco com credenciais de 'user'
conn <- conectabanco("user", "db")

# listando todas as tabelas no banco
dbListTables(conn)

# listando campos (colunas) de uma determinada tabela
dbListFields(conn, "nome_da_tabela")
}

}
