% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{conectabucket}
\alias{conectabucket}
\title{Conexao Com Buckets S3}
\usage{
conectabucket(bucket, prefixo)
}
\arguments{
\item{bucket}{bucket de onde ler os dados}

\item{prefixo}{opcional, prefixo dos arquivos a serem lidos}
}
\value{
objeto de conexao com o arquivamento em bucket S3
}
\description{
Gera a conexao com um mock banco, correspondente a um bucket S3 com arquivos a ler
}
\details{
A implementacao de bancos mock hospedados em buckets funciona exatamente como a de diretorios
locais, descrita em \code{\link{conectalocal}}, incluindo questoes de particionamento.

Este tipo de conexao existe somente para lidar com os casos em que arquivos estao em buckets e
nao no Athena. Neste caso, a conexao deve ser feita como se fosse a um banco relacional comum.
}
