% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{conectalocal}
\alias{conectalocal}
\title{Conexao Com Arquivos Locais}
\usage{
conectalocal(diretorio)
}
\arguments{
\item{diretorio}{diretorio contendo os arquivos csv representando o banco}
}
\value{
objeto de conexao com o arquivamento local
}
\description{
Gera a conexao com um mock banco, correspondente a um diretorio local com arquivos a ler
}
\details{
Os arquivamentos locais suportados por \code{dbrenovaveis} correspondem a diretorios contendo
uma serie de arquivos csv correspondentes as tabelas do banco (algo como um dump de um banco
relacional comum). Os arquivos devem ser nomeados tal qual os nomes de tabelas esperados no banco
assim como os nomes e tipos de dado nas colunas de cada um.

Particionamento de tabelas e suportado, de forma restrita. Inicialmente, para bancos locais
contendo uma ou mais tabelas particionadas, deve existir um arquivo .PARTICAO.json indicando
quais tabelas sao particionadas. Este arquivo tem estrutura muito simples, uma lista de apenas um
nivel com valores booleanos (ver o exemplo em \code{extdata/compart}).

Os dados particionados devem seguir uma estrutura predefinida. Particoes de uma determinada
tabela devem ser nomeadas \code{tabela_partA_partB_...} em que \code{partX} sao chaves
indicadoras apontando os multiplos niveis de particionamento. Por exemplo, particionando-se a
tabela de previstos por usina e modelo, teria-se: previstos_1_1, previstos_2_1, ... previstos_N_1,
previstos_1_2, ..., previstos_N_M, onde N e o maximo id de usinas e M o de modelos.

Nestes casos, a tabela original (previstos, no exemplo acima) deve ser uma tabela mestra contendo
a associacao entre subtabelas e a particao a que correspondem. As tabelas mestras devem conter um
coluna chamada \code{tabela}, e as demais devem ser as colunas pelas quais suas particoes sao
separadas. Voltando ao exemplo de previstos\tabular{lll}{
   tabela \tab id_usina \tab id_modelo \cr
   previstos_1_1 \tab 1 \tab 1 \cr
   previstos_2_1 \tab 2 \tab 1 \cr
   previstos_1_2 \tab 1 \tab 2 \cr
}


A ordem das colunas nao importa, bem como o numero de particoes e colunas pelas quais se
particiona nao precisam necessariamente ser estas do exemplo. As tabelas de particao devem ter
\bold{EXATAMENTE A MESMA ESTRUTURA} da equivalente sem particao, isto e, devem manter as colunas
pelas quais foram originalmente particionadas.

Atualmente apenas particionamentos categoricos sao suportados, isto e, uma faixa de datas nao
funciona como particionamento.
}
