% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{conectamock}
\alias{conectamock}
\title{Conexao Com Mock Bancos}
\usage{
conectamock(schema)
}
\arguments{
\item{schema}{lista contendo o schema do banco, correspondente aos conteudos de um arquivo
\code{schema.json} para banco, ou o caminho de um arquivo deste tipo ou diretorio que o
contenha. Veja Detalhes}
}
\value{
objeto de conexao com o mock banco
}
\description{
Gera a conexao com um mock banco, correspondente a um diretorio local ou s3
}
\details{
Os bancos mock de \code{dbrenovaveis} correspondem a arquivos de dados, em \code{csv} ou \code{parquet}, em
diretorios locais ou buckets no s3 dentro. Para correto funcionamento desta implementacao se
espera uma certa estrutura de arquivos e diretorios de tal modo que o pacote consiga encontrar
os dados relevantes. Recomenda-se ler a vignette
\code{vignette("estrutura-mock", package = "dbrenovaveis")} para maiores detalhes a respeito desta
estrutura de mock banco.

A conexao com bancos mock possui apenas um argumento de entrada: \code{schema}. Este e ou o caminho a
um arquivo json explicitando a estrutura geral do banco ou uma lista contendo o este json ja
lido (a estrutura de schema.json se encontra detalhada na vignette supracitada). O schema de
pode conter uma chave opcional \code{uri}. Ela e opcional pois so seria utilizada no caso das
\code{uri} das tabelas serem caminhos relativos. Caso \code{schema} seja passado como um caminho
e o json lido nao possua essa chave, sera adicionada como igual ao caminho \code{schema}.
}
\examples{

# conexao com o mock banco exemplo do pacote

# passando o diretorio
arq_schema <- system.file("extdata/cpart_parquet/schema.json", package = "dbrenovaveis")
conn1 <- conectamock(arq_schema)

# passando o a lista de schema ja lido
schema <- jsonlite::read_json(arq_schema)
schema$uri <- sub("/schema.json", "", arq_schema)
conn2 <- conectamock(schema)

\dontrun{
identical(conn1, conn2)
}

}
