% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{conectamorgana}
\alias{conectamorgana}
\title{Conexao Com S3 Via Morgana}
\usage{
conectamorgana(schema, x_api_key = Sys.getenv("X_API_KEY"))
}
\arguments{
\item{schema}{lista contendo o schema do banco, correspondente aos conteudos de um arquivo
\code{schema.json} para banco, ou o caminho de um arquivo deste tipo. Veja
\code{\link{conectamock}}}

\item{x_api_key}{chave de api para uso das funcoes que compoem o morgana na aws. Veja Detalhes}
}
\value{
objeto de conexao com o mock banco via morgana
}
\description{
Gera uma conexao mock a um banco no s3, porem realizando queries atraves da engine morgana
}
\details{
A conexao com um banco S3 via morgana e essencialmente a mesma coisa que um mock banco no s3, com
um unico elemento de diferenca sendo a necessidade de uma chave de API para uso das funcoes.

O argumento \code{x_api_key} existe para receber esta chave. Por padrao sera buscada uma variavel
de ambiente \code{"X_API_KEY"} na secao para este argumento. Esta e a abordagem recomendada, de
modo que informacoes pessoais e sensiveis nao ficam expostas hardcoded.

O objeto de saida e, para todos os efeitos, identico a uma conexao mock com o s3. Possui apenas
um atributo adicional que e a chave de api passada originalmente
}
