% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfromdb.r
\name{corrigeposix}
\alias{corrigeposix}
\title{Auxiliar Para Correcao De Datetime}
\usage{
corrigeposix(dat)
}
\arguments{
\item{dat}{data.table com colunas a corrigir corrigir}
}
\value{
Argumento \code{dat} com colunas POSIXct em fuso horario GMT
}
\description{
Corrige o fuso horario de colunas de data em um data.table
}
