% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbrenovaveis.r
\docType{package}
\name{dbrenovaveis}
\alias{dbrenovaveis}
\title{Pacote \code{dbrenovaveis}}
\description{
Acesso Aos Bancos De Dados PEM-Renovaveis
}
\details{
Este pacote contem funcoes para facilitacao do acesso aos
bancos de dados associados a area de energias renovaveis da PEM. Isto
inclui acesso, escrita e resgate de informacoes acerca dos bancos
disponiveis.
}
