% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfromdb.r
\name{decode_output}
\alias{decode_output}
\title{\code{base46} -> \code{data.table}}
\usage{
decode_output(x)
}
\arguments{
\item{x}{vetor de informarcao codificada em base64}
}
\value{
\code{data.table} convertido
}
\description{
Converte resposta de uma consulta a api para dado padronizado
}
