% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.r
\name{funs_check_caminho}
\alias{funs_check_caminho}
\alias{is_abs_path2}
\alias{is_rel_path2}
\title{Checagem De Caminho Absoluto Ou Relativo}
\usage{
is_abs_path2(x)

is_rel_path2(x)
}
\arguments{
\item{x}{vetor de caminhos de arquivos}
}
\value{
vetor booleano indicando se os elementos de \code{x} sao caminhos absolutos ou relativos
dependendo de qual funcao foi chamada
}
\description{
Estas funcoes identificam se um vetor de caminhos \code{x} e absoluto ou nao, incluindo caminhos
no s3
}
