% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{get_ext}
\alias{get_ext}
\title{Extrai Extensao De Arquivo}
\usage{
get_ext(x)
}
\arguments{
\item{x}{nome do arquivo do qual tirar extensao}
}
\value{
extensao do arquivo
}
\description{
Alternativa a \code{\link[tools]{file_ext}} que suporta extensoes com mais de um ponto
}
\details{
\code{\link[tools]{file_ext}} basicamente retorna o ultimo trecho do nome seguindo um ".". Para
arquivos \code{parquet.gzip} isto gera um problema pois acaba retornando apenas \code{gzip}
}
