% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryinfo.r
\name{get_funs_quali}
\alias{get_funs_quali}
\alias{getusinas}
\alias{getmodelos}
\alias{getvertices}
\title{Leitura De Tabelas Qualitativas}
\usage{
getusinas(conexao, usinas, campos = "*")

getmodelos(conexao, modelos, tipo = "previsao", campos = "*")

getvertices(conexao, longitudes = "*", latitudes = "*", campos = "*")
}
\arguments{
\item{conexao}{objeto de conexao ao banco retornado por \code{\link{conectabanco}}}

\item{usinas}{opcional, vetor de strings com codigo das usinas cujas informacoes serao buscadas}

\item{campos}{vetor de strings indicando quais campos (colunas) devem ser lidos}

\item{modelos}{opcional, vetor strings com nome dos modelos cujas informacoes serao buscadas}

\item{tipo}{o tipo de modelo a ser pego: "previsao" para os modelos meteorologicos e "correcao"
para os modelos de correcao de vento}

\item{longitudes}{opcional, vetor de strings com longitudes cujas informacoes serao buscadas}

\item{latitudes}{opcional, vetor de strings com latitudes cujas informacoes serao buscadas}
}
\value{
tabelas qualitativas lidas
}
\description{
Busca informacoes nas tabelas descritivas das informacoes no banco
}
\details{
\code{getusinas} retorna a tabela "usinas" do banco, um dataframe contendo as colunas
especificadas em \code{campos}
}
\seealso{
\code{\link{get_funs_quant}} para funcoes de leitura das tabelas quantitativas
}
