% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfromdb.r
\name{getfromdb}
\alias{getfromdb}
\title{Acesso A Tabelas}
\usage{
getfromdb(conexao, tabela, campos = NA, ...)
}
\arguments{
\item{conexao}{objeto de conexao com um banco de dados}

\item{tabela}{nome da tabela no banco \code{conexao}}

\item{campos}{os campos a reter apos a leitura. Por padrao traz todos}

\item{...}{subsets a serem aplicados. Veja Detalhes e Exemplos}
}
\value{
\code{data.table} dos dados lidos
}
\description{
Funcao generica para leitura de dados em objetos \code{tabela}
}
\details{
Atraves de \code{...} e possivel especificar subsets para a query. Argumentos extras passados por
aqui devem ser nomeados igual ao campo da tabela, ou seu proxy, tomando valor igual ao que deve
ser retido na query.
}
\examples{

# usando banco mock de exemplo do pacote
arq <- system.file("extdata/cpart_parquet/schema.json", package = "dbrenovaveis")
conn <- conectamock(arq)

subbacias <- getfromdb(conn, "subbacias", c("nome", "codigo", "bacia_smap"))

\dontrun{
head(subbacias)
#>             nome     codigo bacia_smap
#> 1: Agua Vermelha  AVERMELHA   GRD_PRNB
#> 2:     B. Bonita    BBONITA  Paranazao
#> 3:  Baixo Iguacu    BAIXOIG        SUL
#> 4:      Capivara   CAPIVARA  Paranazao
#> 5:       Colider    COLIDER      Norte
#> 6:    Emborcacao EMBORCACAO   GRD_PRNB
}

previstos <- getfromdb(conn, "previstos", c("data_previsao", "dia_previsao", "codigo", "qcalc"),
    codigo = "AVERMELHA", dia_previsao = 1, data_previsao = "2020")

\dontrun{
head(previstos)
#>    data_previsao dia_previsao    codigo    qcalc
#> 1:    2020-01-01            1 AVERMELHA 327.1693
#> 2:    2020-01-02            1 AVERMELHA 299.5275
#> 3:    2020-01-03            1 AVERMELHA 283.5529
#> 4:    2020-01-04            1 AVERMELHA 283.1786
#> 5:    2020-01-05            1 AVERMELHA 289.5164
#> 6:    2020-01-06            1 AVERMELHA 326.9825
}

}
