% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{getfromtabela}
\alias{getfromtabela}
\title{Acesso A Tabelas}
\usage{
getfromtabela(tabela, campos = NA, ...)
}
\arguments{
\item{tabela}{objeto da classe \code{tabela} criado por \code{\link{new_tabela}}}

\item{campos}{os campos a reter apos a leitura. Por padrao traz todos}

\item{...}{subsets a serem aplicados. Veja Detalhes e Exemplos}
}
\value{
\code{data.table} dos dados lidos
}
\description{
Funcao generica para leitura de dados em objetos \code{tabela}
}
\details{
Atraves de \code{...} e possivel especificar subsets para a query. Argumentos extras passados por
aqui devem ser nomeados igual ao campo da tabela, ou seu proxy, tomando valor igual ao que deve
ser retido na query.
}
\examples{

# lendo dados verificados
dirloc <- system.file("extdata/sempart", package = "dbrenovaveis")
conect <- conectalocal(dirloc)

# versao simplificada da tabela de usinas
tabusi <- new_tabela(
    nome = "usinas",
    campos = list(
        new_campo("id", "inteiro", FALSE),
        new_campo("codigo", "string", FALSE)),
    conexao = conect)

# representacao da tabela de verificados
tabverif <- new_tabela(
    nome = "verificados",
    campos = list(
        new_campo("id_usina", "inteiro", TRUE, tabusi, "id", "codigo"),
        new_campo("data_hora", "data", FALSE),
        new_campo("vento", "float", FALSE),
        new_campo("geracao", "float", FALSE)
    ),
    conexao = conect)

\dontrun{
# note que nao se especifica "id_usina", mas sim o proxy "codigo"
getfromtabela(tabverif, campos = "geracao", data_hora = "/2021-01-01 10:00:00", codigo = "BAEBAU")
}

}
