% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{le_arquivo_local}
\alias{le_arquivo_local}
\title{Leitor De Arquivo Local}
\usage{
le_arquivo_local(arq)
}
\arguments{
\item{arq}{arquivo a ser lido, em formato csv ou parquet. Deve sempre ser uma tabela regular}
}
\value{
data.table contendo o arquivo lido
}
\description{
Abstracao para leitura de arquivos csv ou parquet
}
