% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-mock.r
\name{le_tabela_mock}
\alias{le_tabela_mock}
\title{Leitor De Tabelas Em Mock Bancos}
\usage{
le_tabela_mock(tabela, arquivo, ...)
}
\arguments{
\item{tabela}{objeto de classe \code{tabela} cujos arquivos componentes devem ser listados}

\item{arquivo}{nome do arquivo componente de \code{tabela} a ser lido, sem extensao}

\item{...}{demais argumentos que possam ser passados para o leitor interno}
}
\value{
data.table contendo a tabela lida
}
\description{
Abstracao para leitura de arquivos csv ou parquet
}
