% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachecreds.r
\name{list_cache_funs}
\alias{list_cache_funs}
\alias{lista_credenciais}
\alias{lista_bancos}
\title{Listagem De Registros Salvos}
\usage{
lista_credenciais()

lista_bancos()
}
\value{
vetor de tags de usuarios ou bancos registrados
}
\description{
Mostra tags dos registros salvos na maquina
}
