% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{lista_conteudo}
\alias{lista_conteudo}
\alias{lista_conteudo.tabela_local}
\alias{lista_conteudo.tabela_s3}
\title{Listagem De Arquivos Por Tabela}
\usage{
lista_conteudo(tabela)

\method{lista_conteudo}{tabela_local}(tabela)

\method{lista_conteudo}{tabela_s3}(tabela)
}
\arguments{
\item{tabela}{objeto de classe \code{tabela} cujos arquivos componentes devem ser listados}
}
\value{
vetor contendo a lista de caminhos completos, local ou s3, dos arquivos que compoem
\code{tabela}
}
\description{
Generica e metodos para listar os arquivos que compoem uma tabela abstrata, local ou no s3
}
