% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{new_campo}
\alias{new_campo}
\title{Construtor Da Classe \code{campo}}
\usage{
new_campo(nome, tipo)
}
\arguments{
\item{nome}{o nome do campo (nome da coluna na tabela)}

\item{tipo}{o tipo de dado contido. Veja Detalhes para os tipos suportados e suas implicacoes}
}
\value{
Objeto da classe \code{campo} contendo especificacao de um campo qualquer
}
\description{
Funcao para geracao de objetos representativos de campos em tabelas num determinado banco
}
\details{
Esta classe compoe o coracao da representacao simbolica de tabelas, utilizada dentro de
\code{\link{new_tabela}} para especificacao das colunas. O argumento \code{tipo} permite detalhar
o tipo de dados contido neste campo, podendo ser um de:

\itemize{
\item{int}
\item{float}
\item{string}
\item{date}
\item{datetime}
}

Cada \code{tipo} de dado gera um tipo especifico de query e possui funcionalidades proprias
durante o processamento de argumentos numa execucao de funcao para formar a query.
}
\seealso{
Construtor de tabelas \code{\link{new_tabela}} para melhor entendimento de seu uso
}
