% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{new_mock}
\alias{new_mock}
\title{Construtor Interno De Mocks}
\usage{
new_mock(schema, morgana = FALSE)
}
\arguments{
\item{schema}{lista contendo o schema do banco, correspondente aos conteudos de um arquivo
\code{schema.json} para banco, ou o caminho de um arquivo deste tipo.}

\item{morgana}{booleano indicando se a conexacao e via morgana ou nao}
}
\value{
objeto da classe \code{mock}
}
\description{
Construtor utilizado para geracao de objetos conexao com mocks e morgana
}
