% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectores.r
\name{parse_inner_reader}
\alias{parse_inner_reader}
\title{Seletor De Inner Reader}
\usage{
parse_inner_reader(extensao)
}
\arguments{
\item{extensao}{extensao dos arquivos no banco}
}
\value{
funcao para leitura dos arquivos
}
\description{
Wrapper para decisao de qual funcao de leitura interna deve ser usada
}
\details{
Atualmente sao suportados apenas os seguintes tipos

\describe{
\item{\code{csv}}{lido atraves de \code{\link[data.table]{fread}}}
\item{\code{parquet|parquet.gzip}}{lido atraves de \code{\link[arrow]{read_parquet}}}
}
}
