% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-mock.r
\name{query2subset}
\alias{query2subset}
\title{Transforma Query Em Expressoes De Subset}
\usage{
query2subset(query)
}
\arguments{
\item{query}{lista contendo termos SELECT, WHERE e ORDERBY de uma query}
}
\value{
mesma lista com os termos em questao convertidos para expressoes de subset
}
\description{
Auxiliar para conversao de strings de query em db para expressoes de subset em data.table
}
