% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{query_local}
\alias{query_local}
\alias{proc_query_local_spart}
\alias{proc_query_local_cpart}
\title{Executores Internos De Query Local}
\usage{
proc_query_local_spart(conexao, query)

proc_query_local_cpart(conexao, query)
}
\arguments{
\item{conexao}{objeto de conexao ao banco retornado por \code{\link{conectabanco}}}

\item{query}{lista detalhando a query como retornado por \code{\link{parseargsOLD}}}
}
\value{
dado recuperado do banco ou erro caso a query nao possa ser realizada
}
\description{
Realizam queries em bancos de dados locais, com ou sem particao
}
