% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{roda_query}
\alias{roda_query}
\title{Executa Queries}
\usage{
roda_query(conexao, query)
}
\arguments{
\item{conexao}{objeto de conexao ao banco retornado por \code{\link{conectabanco}}}

\item{query}{lista detalhando a query como retornado por \code{\link{parseargs}}}
}
\value{
dado recuperado do banco ou erro caso a query nao possa ser realizada
}
\description{
Funcao para execucao de queries nas \code{\link{get_funs_quant}}
}
\details{
Existe um problema na conversao de datas nas viradas de horario de verao, dando erro na query.
Teoricamente isso pode ser resolvido passando tz = "GMT" para as.POSIXlt, porem o DBI nao da
esse nivel de controle.

Para contornar o problema se torna necessario fazer uma maracutaia de trocar o tz do R no momento
da query e retorna-lo ao valor original depois. Ao retornar o valor, o R desloca as datas, entao
e preciso modificar o atributo tzone da coluna de datas de volta para GMT
}
