% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseargs.r
\name{str2mock}
\alias{str2mock}
\title{Converte String De WHERE Para Expressao}
\usage{
str2mock(str)
}
\arguments{
\item{str}{string de subset gerada por uma das funcoes de parse de argumentos}
}
\value{
expressao do subset a ser avaliado
}
\description{
Transforma uma string de WHERE para expressao de subset coerente com \code{data.table}s
}
