% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-leitura.r
\name{switch_reader_func}
\alias{switch_reader_func}
\title{Seleciona Funcao Leitora Por Tipo E Fonte De Arquivo}
\usage{
switch_reader_func(extensao, s3 = FALSE)
}
\arguments{
\item{extensao}{string indicando a extensao de arquivo, preferencialmente precedida de "."}

\item{s3}{booleano indidicano se o arquivo esta no s3 ou nao}
}
\value{
funcao cujo primeiro argumento e o caminho (local ou s3) do arquivo a ser lido e retorna
a leitura executada
}
\description{
Para uma dada \code{extensao} de arquivo e sua origem (local ou s3) retorna uma funcao para
leitura do mesmo
}
\examples{

# para leitura de um json
fun1 <- dbrenovaveis:::switch_reader_func("json", FALSE)
identical(body(fun1)[[1]], jsonlite::read_json)

# para leitura de um csv
fun2 <- dbrenovaveis:::switch_reader_func(".csv", FALSE)
identical(body(fun2)[[1]], data.table::fread)

}
