% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.r
\name{valida_schema_banco}
\alias{valida_schema_banco}
\title{Validador De Schemas De Bancos}
\usage{
valida_schema_banco(schema)
}
\arguments{
\item{schema}{uma lista contendo um \code{schema.json} de banco lido}
}
\value{
lista \code{schema} validada e corrigida, caso necessario, ou erro caso a validacao
encontre problemas
}
\description{
Valida conteudos de um schema de banco e corrige caminhos relativos de tabelas caso necessario
}
