% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.r
\name{valida_schema_tabela}
\alias{valida_schema_tabela}
\title{Validador De Schemas De Tabelas}
\usage{
valida_schema_tabela(schema)
}
\arguments{
\item{schema}{uma lista contendo um \code{schema.json} de tabela lido}
}
\value{
lista \code{schema} validada ou erro, caso a validacao encontre problemas
}
\description{
Valida conteudos de um \code{schema.json} de tabelas
}
