% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-funs.r
\name{cache_funs}
\alias{cache_funs}
\alias{registra_credenciais}
\alias{registra_banco}
\title{Funcoes Para Caching De Informacoes}
\usage{
registra_credenciais(usuario, senha, tag)

registra_banco(host, port, database, tag)
}
\arguments{
\item{usuario}{string contendo o nome de usuario utilizado no acesso ao banco}

\item{senha}{string contendo a senha utilizada no acesso ao banco}

\item{tag}{um identificador do conjunto de informacoes sendo registradas}

\item{host}{string contendo o endereco de ip no qual o banco esta hospedado}

\item{port}{string contendo a porta para acesso}

\item{database}{string contendo o nome do banco a ser acessado}
}
\value{
salva as informacoes como arquivo binario em um diretorio do pacote
}
\description{
Registro de informacoes tanto dos bancos quanto de credenciais de usuario
}
\examples{

\dontrun{
# registrando o usuario 'jose' sob o mnemonico
registra_credenciais("jose", "senha", tag = "usuario_jose")

# registrando banco que "jose" vai acessar
registra_banco("hopedagem", 0000, "nome_do_banco", tag = "banco_do_jose")

# com isso, e possivel conectar-se
conn <- conectabanco("usuario_jose", "banco_do_jose")
}

}
\seealso{
\code{\link{lista_credenciais}} \code{\link{lista_bancos}} para buscar quais registros
existem salvos na maquina
}
