% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryinfo.r
\name{get_funs_quali}
\alias{get_funs_quali}
\alias{getusinas}
\alias{getmodelosprev}
\title{Leitura De Tabelas Qualitativas}
\usage{
getusinas(conexao, campos)

getmodelosprev(conexao, campos)
}
\arguments{
\item{conexao}{objeto de conexao ao banco retornado por \code{\link{conectabanco}}}

\item{campos}{vetor de strings indicando quais campos (colunas) devem ser lidos}
}
\value{
tabelas qualitativas lidas
}
\description{
Busca informacoes nas tabelas descritivas das informacoes no banco
}
\details{
\code{getusinas} retorna a tabela "usinas" do banco, um dataframe contendo as colunas
especificadas em \code{campos}
}
\seealso{
\code{\link{get_funs_quant}} para funcoes de leitura das tabelas quantitativas
}
