% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{le_tabela_mock}
\alias{le_tabela_mock}
\title{Leitor De Tabelas Em Mock Bancos}
\usage{
le_tabela_mock(conexao, tabela, ...)
}
\arguments{
\item{conexao}{objeto de conexao ao banco retornado por \code{\link{conectabanco}}}

\item{tabela}{tabela a ser lida, em formato csv ou parquet. Deve sempre ser uma tabela regular}

\item{...}{demais argumentos que possam ser passados para o leitor interno}
}
\value{
data.table contendo a tabela lida
}
\description{
Abstracao para leitura de arquivos csv ou parquet
}
