% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{new_campo}
\alias{new_campo}
\title{Construtor Da Classe \code{campo}}
\usage{
new_campo(nome, tipo)
}
\arguments{
\item{nome}{o nome do campo (nome da coluna na tabela)}

\item{tipo}{o tipo de dado contido. Veja Detalhes para os tipos suportados e suas implicacoes}
}
\value{
Objeto da classe \code{campo} contendo especificacao de um campo qualquer
}
\description{
Funcao para geracao de objetos representativos de campos em tabelas num determinado banco
}
\details{
Esta classe compoe o coracao da representacao simbolica de tabelas, utilizada dentro de
\code{\link{new_tabela}} para especificacao das colunas. O argumento \code{tipo} permite detalhar
o tipo de dados contido neste campo, podendo ser um de:

\itemize{
\item{inteiro}
\item{string}
\item{float}
\item{data}
}

Cada tipo de dado gera um tipo especifico de query e possui funcionalidades proprias durante o
processamento de argumentos numa execucao de funcao para formar a query.
}
\examples{

# Retomando os exemplos de 'new_tabela'

dirloc <- system.file("extdata/sempart", package = "dbrenovaveis")
conect <- conectalocal(dirloc)

tabusi <- new_tabela(
    nome = "usinas",
    campos = list(
        new_campo("id", "inteiro"),
        new_campo("codigo", "string")),
    conexao = conect)

# Similar ao exemplo da descricao, vemos que a tabela de verificados nao possui uma coluna 
# 'codigo' com a qual pesquisar por usinas, apenas o id inteiro arbitrario.
# Ao especificar o campo 'id_usina' com proxy 'codigo' na tabela 'tabusi', o pacote sabera que
# uma query sobre usinas na tabela de verificados pode ser feita na coluna 'codigo' de 'tabusi'

tabverif <- new_tabela(
    nome = "verificados",
    campos = list(
        new_campo("id_usina", "inteiro"),
        new_campo("data_hora", "data"),
        new_campo("vento", "float"),
        new_campo("geracao", "float")
    ),
    conexao = conect)

}
