% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseargs.r
\name{parsearg}
\alias{parsearg}
\title{Generica Para Parse De Argumentos Unicos}
\usage{
parsearg(campo, valor, ...)
}
\arguments{
\item{campo}{objeto da classe \code{campo} gerado por \code{\link{new_campo}}}

\item{valor}{o valor buscado na tabela para montar a query}

\item{...}{demais argumentos que possam existir nos metodos de cada tipo de dado}
}
\value{
string contendo o trecho de WHERE relacionado ao campo em questao
}
\description{
Funcao generica dos metodos de parse de argmumentos passados nas queries de tabela
}
\details{
\code{parseargsOLD} e uma funcao generica para processar argumentos passados nas funcoes de query.
Cada tipo de campo (data, inteiro, string e etc.) possui um metodo proprio que retorna uma parte
do elemento WHERE da query associada ao campo passado.

No caso de campos do tipo \code{campo_data}, \code{valor} deve ser uma string indicando a faixa
de datahoras no formato \code{"YYYY[-MM-DD HH:MM:SS]/YYYY[-MM-DD HH:MM:SS]"}. As partes entre
colchetes sao opcionais, tanto no limite inicial quanto final.
}
\examples{

tabusi <- new_tabela(
    nome = "usinas",
    campos = list(
        new_campo("id", "inteiro", FALSE),
        new_campo("codigo", "string", FALSE)),
    conexao = conect)

# parse num campo sem proxy
\dontrun{
parsearg(tabusi$campos[[1]], codigo = "RNUEM3")
}

\dontrun{
# parse num campo com proxy
campo <- new_campo("id_usina", "inteiro", TRUE, tabusi, "id", "codigo")
parsearg(campo, valor = "BAEBAU")
}

\dontrun{
# parse de datas
campo <- new_campo("datahora", "data")
parsearg(campo, valor = "2020-01-01/2021-02-13 13:40")
}

}
