% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseargs.r
\name{parseargs}
\alias{parseargs}
\title{Wrapper De Parses Para Queries}
\usage{
parseargs(tabela, campos = NA, ...)
}
\arguments{
\item{tabela}{objeto da classe \code{tabela} criado por \code{\link{new_tabela}}}

\item{campos}{os campos a reter apos a leitura. Por padrao traz todos}

\item{...}{subsets a serem aplicados. Veja Detalhes e Exemplos}
}
\value{
lista contendo os trechos de query para executar no banco
}
\description{
Interpreta argumentos para montar a string da query a ser executada no banco
}
\seealso{
generica para parse de cada campo individualmente \code{\link{parsearg}};
}
