% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querydados.r
\name{parsedatas}
\alias{parsedatas}
\title{Parse Argumento \code{datahoras}}
\usage{
parsedatas(datahoras, nome)
}
\arguments{
\item{datahoras}{uma string indicando faixa de tempo, como descrito em \code{\link{get_funs}}}

\item{nome}{nome do campo para query}
}
\value{
string contendo a condicao de busca associada a datas na query
}
\description{
Transforma \code{datahoras} passado as funcoes \code{get*} numa string de query
}
\details{
Esta funcao e responsavel por traduzir as strings de janela temporal \code{datahora} das funcoes
\code{\link{get_funs}} em condicionais de WHERE para a query.

Por padrao, todas as queries serao realizadas no padrao \code{data >= lim_1 AND data < lim_2},
isto e, serao sempre buscados intervalos fechados no inicio e abertos no final. Desta forma todos
os padroes de janela podem ser representados da mesma forma, simplificando a funcao.
}
\examples{

cond <- parsedatas("2021", "data_hora")
\dontrun{
identical(cond, "data_hora >= '2021-01-01 00:00:00' AND data_hora < '2022-01-01 00:00:00'")
}

cond <- parsedatas("2020-11-30 12:30", "data_hora")
\dontrun{
identical(cond, "data_hora >= '2020-11-30 12:30:00' AND data_hora < '2020-11-30 12:30:01'")
}

}
