% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{schema2tabela}
\alias{schema2tabela}
\title{Construtor Externo}
\usage{
schema2tabela(schema)
}
\arguments{
\item{schema}{ou uma lista de schema.json já lido ou o caminho do arquivo}
}
\value{
objeto \code{tabela}
}
\description{
Gera objetos \code{tabela} a partir de um schema.json. Wrapper de \code{new_tabela}
}
\examples{

# Usando as mesmas tabelas exemplificadas na doc de `new_tabela`
arq1 <- system.file("extdata/cpart_parquet/subbacias/schema.json", package = "dbrenovaveis")
tab1 <- dbrenovaveis:::schema2tabela(arq1)
arq2 <- system.file("extdata/cpart_parquet/vazoes/schema.json", package = "dbrenovaveis")
tab2 <- dbrenovaveis:::schema2tabela(arq2)

}
